Adopt commit 02108f162 from master: Merge pull request #1138 from hrosen4/conversion_drqgs-100755799
commit 02108f162e26b72f4828a38bd81047c680806d69
Merge: c11b22604 d50845af2
Author: Henry Verschell <hverschell@bloomberg.net>
Date:   Mon May 15 17:09:44 2017 -0400

    Merge pull request #1138 from hrosen4/conversion_drqgs-100755799

    Conversion warning

commit d50845af2995ed4a9b9f2c2a2647b83111e2e45d
Author: Hyman Rosen <hrosen4@bloomberg.net>
Date:   Mon May 15 16:54:29 2017 -0400

    Conversion warning

+++ groups/bal/balber/balber_berutil.h
@@ -929,7 +929,7 @@ int BerUtil_Imp::numBytesToStream(TYPE value)
             // wire.  The leading zero is followed by all of the bytes of the
             // unsigned value.
 
-            return sizeof(TYPE) + 1;                                  // RETURN
+            return static_cast<int>(sizeof(TYPE) + 1);                // RETURN
         }
 
         // This mask zeroes out the most significant byte and the first bit of
